Add preparing-dessert tag in buildings.xml
08/31/2022

- rework: make preparing dessert function independent from
          cooking function
- improve: add one preparing dessert's activity spot in 3 buildings
- rework: pinpoint new cooking and food production activity spots
          for building 'lounge'
- style: Update furniture and activity spot locations in lounge.svg

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/robot/ai/job/Chefbot.java
@@ -84,7 +84,7 @@ implements Serializable {
 
 		// Add total population / 10.
 		int population = settlement.getIndoorPeopleCount();
-		result+= ((double) population / 10D);
+		result+= population / 10D;
 
 		return result;			
 	}
